make it all fit in ram

+++ stm/src/rng.h
@@ -6,7 +6,7 @@
 namespace rng {
 	void init();
 	uint32_t get();
-	uint8_t getclr();
+	uint16_t getclr();
 
 	inline static float getrange(float min, float max) {
 		uint32_t prec = (max - min) * 1000;
